Extracts screenshot test drivers
